requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=1.0.0
python-docx>=0.8.11
weasyprint>=60.2
google-generativeai>=0.3.1
python-dotenv>=1.0.0
pydantic>=2.0.0
click>=8.1.7